% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcr_abl1.R
\name{bcr_abl1}
\alias{bcr_abl1}
\title{Function to extract BCR::ABL1 principal genetic abnormality}
\usage{
bcr_abl1(person, measurement)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}
}
\value{
A data frame containing the BCR::ABL1 principal genetic abnormality
}
\description{
Function to extract BCR::ABL1 principal genetic abnormality
}
