% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_karyotype.R
\name{complex_karyotype}
\alias{complex_karyotype}
\title{Function to extract complex karyotype principal genetic abnormality}
\usage{
complex_karyotype(person, measurement)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}
}
\value{
A data frame containing the complex karyotype principal genetic abnormality
}
\description{
Function to extract complex karyotype principal genetic abnormality
}
