% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etv6_runx1.R
\name{etv6_runx1}
\alias{etv6_runx1}
\title{Function to extract the ETV6::RUNX1 principal genetic abnormality}
\usage{
etv6_runx1(person, measurement)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}
}
\value{
A data frame containing the ETV6::RUNX1 principal genetic abnormality
}
\description{
Function to extract the ETV6::RUNX1 principal genetic abnormality
}
