% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_B_other_B_other_plus_and_T_other.R
\name{extract_B_other_B_other_plus_and_T_other}
\alias{extract_B_other_B_other_plus_and_T_other}
\title{Function to extract B-others and B-other-plus principal genetic abnormality}
\usage{
extract_B_other_B_other_plus_and_T_other(
  person,
  measurement,
  condition_occurrence,
  visit_occurrence
)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}

\item{condition_occurrence}{The condition_occurrence data frame from the OMOP table}
}
\value{
A data frame containing the B-others and B-other-plus principal genetic abnormality
}
\description{
Function to extract B-others and B-other-plus principal genetic abnormality
}
