% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_complex_karyotype.R
\name{extract_complex_karyotype}
\alias{extract_complex_karyotype}
\title{Function to extract complex karyotype principal genetic abnormality}
\usage{
extract_complex_karyotype(person, measurement, visit_occurrence)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}

\item{visit_occurrence}{The visit_occurrence data frame from the OMOP table}
}
\value{
A data frame containing the complex karyotype principal genetic abnormality
}
\description{
Function to extract complex karyotype principal genetic abnormality
}
