% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_etv6_runx1.R
\name{extract_etv6_runx1}
\alias{extract_etv6_runx1}
\title{Function to extract the ETV6::RUNX1 principal genetic abnormality}
\usage{
extract_etv6_runx1(person, measurement, visit_occurrence)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}

\item{visit_occurrence}{The visit_occurrence data frame from the OMOP table}
}
\value{
A data frame containing the ETV6::RUNX1 principal genetic abnormality
}
\description{
Function to extract the ETV6::RUNX1 principal genetic abnormality
}
