% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_karyotype.R
\name{extract_karyotype}
\alias{extract_karyotype}
\title{Function to extract karyotype from the OMOP measurement table}
\usage{
extract_karyotype(measurement)
}
\arguments{
\item{measurement}{The measurement data frame from the OMOP table}
}
\value{
The extracted karyotype in a data frame format
}
\description{
Function to extract karyotype from the OMOP measurement table
}
