% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_kmt2a_mllt1.R
\name{extract_kmt2a_mllt1}
\alias{extract_kmt2a_mllt1}
\title{Function to extract KMT2A::MLLT1 principal genetic abnormality}
\usage{
extract_kmt2a_mllt1(person, measurement, visit_occurrence)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}

\item{visit_occurrence}{The visit_occurrence data frame from the OMOP table}
}
\value{
A data frame containing the KMT2A::MLLT1 principal genetic abnormality
}
\description{
Function to extract KMT2A::MLLT1 principal genetic abnormality
}
