% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_kmt2a_r.R
\name{extract_kmt2a_r}
\alias{extract_kmt2a_r}
\title{Function to extract KMT2A_r principal genetic abnormality}
\usage{
extract_kmt2a_r(person, measurement)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}
}
\value{
A data frame containing the KMT2A_r principal genetic abnormality
}
\description{
Function to extract KMT2A_r principal genetic abnormality
}
