% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tcf3_hlf.R
\name{extract_tcf3_hlf}
\alias{extract_tcf3_hlf}
\title{Function to extract TCF3::HLF principal genetic abnormality}
\usage{
extract_tcf3_hlf(person, measurement)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}
}
\value{
A data frame containing the TCF3::HLF principal genetic abnormality
}
\description{
Function to extract TCF3::HLF principal genetic abnormality
}
