% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tcf3_pbx1.R
\name{extract_tcf3_pbx1}
\alias{extract_tcf3_pbx1}
\title{Function to extract TCF3::PBX1 principal genetic abnormality}
\usage{
extract_tcf3_pbx1(person, measurement, visit_occurrence)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}

\item{visit_occurrence}{The visit_occurrence data frame from the OMOP table}
}
\value{
A data frame containing the TCF3::PBX1 principal genetic abnormality
}
\description{
Function to extract TCF3::PBX1 principal genetic abnormality
}
