% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsubtype_and_riskgroup.R
\name{gsubtype_and_riskgroup}
\alias{gsubtype_and_riskgroup}
\title{Function to extract gsubtype and riskgroup principal genetic abnormality}
\usage{
gsubtype_and_riskgroup(person, measurement, condition_occurrence)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}

\item{condition_occurrence}{The condition_occurrence data frame from the OMOP table}
}
\value{
A data frame containing the gsubtype and riskgroup principal genetic abnormality
}
\description{
Function to extract gsubtype and riskgroup principal genetic abnormality
}
