% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heh_hap_hotr.R
\name{heh_hap_hotr}
\alias{heh_hap_hotr}
\title{Function to extract ANEUOPLOIDY principal genetic abnormality}
\usage{
heh_hap_hotr(person, measurement)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}
}
\value{
A data frame containing the ANEUOPLOIDY principal genetic abnormality
}
\description{
Function to extract ANEUOPLOIDY principal genetic abnormality
}
