% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamp21.R
\name{iamp21}
\alias{iamp21}
\title{Function to extract IAMP21 principal genetic abnormality}
\usage{
iamp21(person, measurement)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}
}
\value{
A data frame containing the IAMP21 principal genetic abnormality
}
\description{
Function to extract IAMP21 principal genetic abnormality
}
