% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kary_status.R
\name{kary_status}
\alias{kary_status}
\title{Function to extract karyotype status from the OMOP measurement table}
\usage{
kary_status(person, measurement)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}
}
\value{
A data frame containing the karyotype status
}
\description{
Function to extract karyotype status from the OMOP measurement table
}
