% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmt2a_mllt1.R
\name{kmt2a_mllt1}
\alias{kmt2a_mllt1}
\title{Function to extract KMT2A::MLLT1 principal genetic abnormality}
\usage{
kmt2a_mllt1(person, measurement)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}
}
\value{
A data frame containing the KMT2A::MLLT1 principal genetic abnormality
}
\description{
Function to extract KMT2A::MLLT1 principal genetic abnormality
}
