% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_abn_B_other_B_other_plus.R
\name{principal_abn_B_other_B_other_plus}
\alias{principal_abn_B_other_B_other_plus}
\title{Function to extract B-others and B-other-plus principal genetic abnormality}
\usage{
principal_abn_B_other_B_other_plus(person, measurement, condition_occurrence)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}

\item{condition_occurrence}{The condition_occurrence data frame from the OMOP table}
}
\value{
A data frame containing the B-others and B-other-plus principal genetic abnormality
}
\description{
Function to extract B-others and B-other-plus principal genetic abnormality
}
