% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_abn_T_other_no_data.R
\name{principal_abn_T_other_no_data}
\alias{principal_abn_T_other_no_data}
\title{Function to extract T-other and No data principal genetic abnormality}
\usage{
principal_abn_T_other_no_data(person, measurement, condition_occurrence)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}

\item{condition_occurrence}{The condition_occurrence data frame from the OMOP table}
}
\value{
A data frame containing the T-other and No data principal genetic abnormality
}
\description{
Function to extract T-other and No data principal genetic abnormality
}
