% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_iamp21.R
\name{extract_iamp21}
\alias{extract_iamp21}
\title{Function to extract IAMP21 principal genetic abnormality}
\usage{
extract_iamp21(person, measurement, visit_occurrence)
}
\arguments{
\item{person}{The person data frame from the OMOP table}

\item{measurement}{The measurement data frame from the OMOP table}
}
\value{
A data frame containing the IAMP21 principal genetic abnormality
}
\description{
Function to extract IAMP21 principal genetic abnormality
}
